requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.18.1
webdriver-manager==4.0.1
PyPDF2==3.0.1
pdfkit==1.0.0
html2pdf==0.1
PyYAML==6.0.1 